ai_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
ultralytics==8.2.0
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.26.4
